Iteration on copyright_header.py to use regexes and subtree-specific rules
The validation of the copyright header is done with a more comprehensive
regex check for the exact format.

The output of the 'report' command has changed to give a per-subtree
breakdown.

Also implements 'check' command to give non-zero status and hints on how
to fix up files that don't have acceptable copyright headers.