Iteration on copyright_header.py to use regexes and add 'check' subcommand
The validation of the copyright header is done with a more comprehensive regex
check for the exact format.

The output of the 'report' command has changed to give the results of specific
expectations for sets of files.

Also implements 'check' command to give non-zero shell status if issues are
found along with hints on what needs to be done to resolve the issue.